type disk manager
